Move BuildScenario to top-level class
